연도, 월 별로 구매목록을 불러오도록 하였습니다. 1. next 파일 라우팅을 이용하여 기존 paymentList파일의 도메인을 localhost:3000/paymentList로 기본 도메인 설정함 (paymentList/index.js가 기존의 paymentList) 2. paymentList/[param]에서 param에 따라 api를 호출하도록 함 (ex) paymentList?year=2024&month=1 일 경우, 2024년, 1월달에 해당하는 구매목록 리스트 불러옴) 3. 기존에 DB에서 date의 타입이 String 이었던 점을 수정하여 Date 타입으로 변경하였으며, 이에 따라 domain도 수정함 4.  paymentList?year=2024&month=1 와 같은 방식을 통해 해당하는 연도, 월에 해당하는 구매목록 리스트를  불러온 후 상단의 '구매목록' 버튼을 눌렀을 시 전체 구매목록이 불러오지지 않았던 점 수정 (기존에 상단 바가 Link였던 것을 a 태그를 이용하였음
