+++ AutumnShop/front/Autumnshop/pages/paymentList/index.js
@@ -1,0 +1,4 @@
+// pages/paymentList/index.js
+import PaymentList from "./[param]";
+
+export default PaymentList;
