+++ AutumnShop/src/main/java/com/example/AutumnMall/repository/PaymentRepository.java
@@ -3,10 +3,11 @@ package com.example.AutumnMall.repository;
 import com.example.AutumnMall.domain.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.time.LocalDate;
 import java.util.List;
-import java.util.Optional;
 
 public interface PaymentRepository extends JpaRepository<Payment, Long> {
     List<Payment> findByMemberId(Long memberId);
 
+    List<Payment> findByMemberIdAndDateBetween(Long memberId, LocalDate startDate, LocalDate endDate);
 }
